deleted subfolder Contact in folder Contact, moved Burger folder to Header folder, fixed conlict with button css in Tracking section
