--- src/components/Contact/Burger/Backdrop.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import './Backdrop.scss';
-
-const Backdrop = props => (
-    <div className="backdrop" onClick={props.click}/>
-)
-
-export default Backdrop;
