--- src/components/Contact/Burger/Backdrop.scss
@@ -1,9 +1,0 @@
-.backdrop {
-    background: rgba(0,0,0,0.3);
-    height: 100%;
-    left: 0;
-    position: fixed;
-    top: 0;
-    width: 100%;
-    z-index: 100;
-}
