+++ src/components/Header/Burger/GeneralBurger.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import Navbar from './Navbar';
 import SideNav from './SideNav';
-import Backdrop from './Backdrop';
 
 class GeneralBurger extends React.Component {
   state = {
@@ -25,7 +24,6 @@ class GeneralBurger extends React.Component {
 
     if (this.state.sideDrawerOpen) {
       sideDrawer = <SideNav />
-      backdrop = <Backdrop click={this.backdropClickHandler}/>
     }
     return (
       <div style={{height: '100%'}}>
