+++ src/components/TrackingSection/TrackingSection.scss
@@ -80,7 +80,7 @@ body {
 
 @media (min-width: 576px) and (max-width: 800px) {
     
-    Button {
+    .tracking_button {
            font-size: 95.7%!important;
            position: relative;
            right:-43.5%;
@@ -110,7 +110,7 @@ body {
     
 
 @media (max-width: 575px) {
-    Button {
+    .tracking_button {
       position: relative;
       top: 45px;
       left: 0px;
