--- src/components/header/Button/Button.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import './Button.scss';
-
-
-
-const Button = ({title}) => (
-     <div className="div-btn">
-          <button className="login-button">{title}</button>        
-     </div>
-     
- 
-)
-
-export default Button;
