--- src/components/header/Button/Button.scss
@@ -1,25 +1,0 @@
-.div-btn{
-   text-align: center;
-   }
-   
-   .login-button{
-       background-color: #C9A959;
-       color:white;
-       border-radius: 2px;
-       cursor: pointer;
-       box-shadow: 0 0.1875rem .1875rem 0 rgba(0, 0, 0,.1);
-       padding: 1.25rem 2rem;
-       font-family: 'Varela Round', sans-serif;
-       font-size: 80%;
-       line-height: 1.5;
-       text-transform: uppercase;
-       border-radius: .25rem;
-       letter-spacing: .15rem;
-       border: 0;
-       font-weight: 400;
-      }
-   
-    
-    .login-button:hover{
-       background-color: #b99a6f;
-    }
