--- src/components/header/Header.scss
@@ -1,82 +1,0 @@
-
-.header_header{
-    background-color: transparent; 
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-    min-height: 56px;
-    padding: 8px 44px;
-    position: absolute;
-    position: fixed;
-    width: 100%;
-    z-index: 1;    
-}
-
-.header_header_white{
-    background-color: white; 
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-    min-height: 56px;
-    padding: 8px 44px;
-    position: absolute;
-    position: fixed;
-    width: 100%;
-    z-index: 1;   
-}
-
-.header_link{
-    color: white;
-    font-family: nunito;
-    font-size: 16px;
-    letter-spacing: .0625em;
-    text-decoration: none;
-}
-
-.header_link:hover {
-    color: fade(white, 80%);
-    outline: none;
-    background-color: transparent;
-}
-
-.header_link_black{
-    color: black;
-    font-family: nunito;
-    font-size: 16px;
-    letter-spacing: .0625em;
-    text-decoration: none;
-}
-
-.header_link_black:hover{
-    color: #C9A959;
-
-}
-
-.header_link_div{
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-    width: 100%;
-    justify-content: flex-end;
-}
-
-.selected{
-    border-bottom: 0.25rem solid #C9A959;
-}
-
-.burger-btn {
-  display: none;
-}
-
-@media (max-width: 980px) {
-    .header_link_div {
-        display: none;
-    }
-    // .header_header {
-    //     background-color: white;
-    // }
-    .burger-btn {
-        display: contents;
-    }
-
-}
