--- src/components/header/PageHeader/PageHeader.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import Button from '../Button/Button';
-
-import './PageHeader.scss'
-
-
-const PageHeader = props => (
-<>
-    <header className="masthead">
-
-        <h1>Yoship Experience Delivery</h1>
-        <h2>A new way of delivery for luxury brands and demanding clients.
-        This is not a simple package, this is an YoShip experience.</h2>
-        <Button title= "GET STARTED"/>
-
-    </header>
-</>
-
-)
-export default PageHeader;
