--- src/components/header/PageHeader/PageHeader.scss
@@ -1,64 +1,0 @@
-
-
-*, ::after, ::before {
-  box-sizing: border-box;
-}
-.masthead {
-  position: relative;
-  width: 100%;
-  height: 100vh;
-  min-height: 35rem;
-  padding: 15rem 0;
-  
-
-
-  background: -webkit-gradient(linear, left top, left bottom, from(rgba(22, 22, 22, 0.9)), color-stop(75%, rgba(22, 22, 22, 0.7)), to(#161616)), url("../../../assets/images/header-background.jpg");
-  background: linear-gradient(to bottom, rgba(22, 22, 22, 0.7) 0%, rgba(22, 22, 22, 0.7) 90%, #161616 100%), url("../../../assets/images/header-background.jpg");
-  background-position: center;
-  background-repeat: no-repeat;
-  background-attachment: scroll;
-  background-size: cover;
-}
-
-.masthead h1 {
-  font-family: 'Varela Round', 'sans-serif';
-  font-size: 2.0rem;
-  line-height: 2.5rem;
-  font-weight: 100;
-  letter-spacing: 0.5rem;
-  background: -webkit-linear-gradient(rgb(255, 255, 255), rgba(255, 194, 136, 0.4));
-  -webkit-text-fill-color: transparent;
-  -webkit-background-clip: text;
-  text-align: center;
-}
-
-.masthead h2 {
-  font-family: 'Varela Round', 'sans-serif';
-  max-width: 20rem;
-  font-size: 1rem;
-  font-weight: 100;
-  line-height: 1.2;
-  color: rgba(255, 255, 255, 0.7);
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: 3rem;
-  margin-top: .5rem;
-  text-align: center;
-}
-
-@media (min-width: 980px) {
-  
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
