--- src/components/header/PageHeader/titlesBtn.json
@@ -1,1 +1,0 @@
-{"headerPage": "GET STARTED"}
