--- src/components/header/Tab/Tab.scss
@@ -1,11 +1,0 @@
-.header_tab{
-    padding: 2rem 1.5rem;
-}
-
-@media (max-width: 980px) {
-    .header_tab {
-        font-size: 1rem;
-        margin: -25px;
-        text-align: left;
-    }
-}
