+++ src/screen/Home.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PageHeader from '../components/Header/PageHeader/PageHeader';
 import ProjectSection from '../components/ProjectSection/ProjectSection';
-import Contact from '../components/Contact/Contact/Contact';
+import Contact from '../components/Contact/Contact';
 import Footer from '../components/Footer/Footer'
 import TrackingSection from "../components/TrackingSection/TrackingSection";
